+++ pyvoronoi/voronoi.cpp
@@ -19,7 +19,7 @@ void VoronoiDiagram::Construct() {
 
 void VoronoiDiagram::GetEdges(std::vector<c_Vertex> &vertices, std::vector<c_Edge> &edges) {
 	std::set<const voronoi_edge<double> *> visited;
-	std::map<const voronoi_vertex<double> *, size_t> vertexMap;
+	std::map<const voronoi_vertex<double> *, long long> vertexMap;
 
 	for (voronoi_diagram<double>::const_edge_iterator it = vd.edges().begin(); 	it != vd.edges().end(); ++it) {
 		if (visited.find(&(*it)) != visited.end())
@@ -28,14 +28,14 @@ void VoronoiDiagram::GetEdges(std::vector<c_Vertex> &vertices, std::vector<c_Edg
 		const voronoi_vertex<double> *start = it->vertex0();
 		const voronoi_vertex<double> *end = it->vertex1();
 		
-		size_t startIndex = -1;
+		long long startIndex = -1;
 		if (start != 0) {
-			std::map<const voronoi_vertex<double> *, size_t>::iterator it = vertexMap.find(start);
+			std::map<const voronoi_vertex<double> *, long long>::iterator it = vertexMap.find(start);
 
 			if (it == vertexMap.end())
 			{
 				c_Vertex startVertex = c_Vertex(start->x(), start->y());
-				startIndex = vertices.size();
+				startIndex = (long long) vertices.size();
 				vertices.push_back(startVertex);
 				vertexMap[start] = startIndex;
 			}
@@ -44,14 +44,14 @@ void VoronoiDiagram::GetEdges(std::vector<c_Vertex> &vertices, std::vector<c_Edg
 			}
 		}
 		
-		size_t endIndex = -1;
+		long long endIndex = -1;
 		if (end != 0) {
-			std::map<const voronoi_vertex<double> *, size_t>::iterator it = vertexMap.find(end);
+			std::map<const voronoi_vertex<double> *, long long>::iterator it = vertexMap.find(end);
 
 			if (it == vertexMap.end())
 			{
 				c_Vertex endVertex = c_Vertex(end->x(), end->y());
-				endIndex = vertices.size();
+				endIndex = (long long) vertices.size();
 				vertices.push_back(endVertex);
 				vertexMap[end] = endIndex;
 			}
